+++ tool/travis.sh
@@ -5,3 +5,11 @@ set -e
 
 # analyze, test, check formatting
 pub run grinder bot
+
+if [ "$COVERALLS_TOKEN" ]; then
+  pub global activate dart_coveralls
+  pub global run dart_coveralls report \
+    --retry 2 \
+    --exclude-test-files \
+    test/homedir_test.dart
+fi