+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log for homedir
 Get home directory on any platform
 
+## [Unreleased](http://github.com/kasperpeulen/homedir.dart/compare/0.0.4...HEAD)
+
+### Docs
+* improve README.md [e7bf532](https://github.com/kasperpeulen/homedir.dart/commit/e7bf532fb7cc8845f1d14b0633ac65029ae809de)
+* make the travis link right [273d6b5](https://github.com/kasperpeulen/homedir.dart/commit/273d6b5c61268e56e68beaec14e4d0004631bfb3)
+
 ## [0.0.4](http://github.com/kasperpeulen/homedir.dart/compare/0.0.3...0.0.4) - 2015-12-11
 
 ### Docs